+++ mrv2/python/demos/fltk/bad/sys_exit_test.py
@@ -21,14 +21,14 @@
 # Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307
 # USA.
 #
-# Please report all bugs and problems to "pyfltk-user@lists.sourceforge.net".
+# Please report all bugs and problems to "pyfltk14-user@lists.sourceforge.net".
 #
 
 
 import sys, os
-import fltk
+import fltk14
 
-class mybutton(fltk.Fl_Button):
+class mybutton(fltk14.Fl_Button):
     def __init__(self, x,y,w,h,label):
         super().__init__(x,y,w,h, label)
         self.callback(self.foo)
@@ -38,7 +38,7 @@ class mybutton(fltk.Fl_Button):
         print("Button pressed: exit!")
         sys.exit()
 
-class subwindow(fltk.Fl_Gl_Window):
+class subwindow(fltk14.Fl_Gl_Window):
     global window
     def __init__(self, x,y,w,h):
         super().__init__(x,y,w,h, "")
@@ -47,14 +47,14 @@ class subwindow(fltk.Fl_Gl_Window):
         pass
 
     def handle(self, event):
-        if event == fltk.FL_PUSH:
+        if event == fltk14.FL_PUSH:
             print("Push event detected: exit!")
             sys.exit()
 
             
         return super().handle(event)
 
-window = fltk.Fl_Double_Window(0, 0, 800, 600, "xxx")
+window = fltk14.Fl_Double_Window(0, 0, 800, 600, "xxx")
 
 sw = subwindow(0,0,800,500)
 window.resizable(sw)
@@ -64,5 +64,5 @@ mb = mybutton(350,520,100,60, "Exit")
 window.end()
 window.show()
 
-fltk.
+fltk14.